+++ src/main/scala/com/dataintuitive/viash/functionality/dataobjects/DoubleObject.scala
@@ -26,6 +26,8 @@ case class DoubleObject(
   example: OneOrMore[Double] = Nil,
   default: OneOrMore[Double] = Nil,
   required: Boolean = false,
+  min: OneOrMore[Double] = Nil,
+  max: OneOrMore[Double] = Nil,
   direction: Direction = Input,
   multiple: Boolean = false,
   multiple_sep: Char = ':',
@@ -43,6 +45,6 @@ case class DoubleObject(
     multiple: Boolean,
     multiple_sep: Char
   ): DataObject[Double] = {
-    copy(name, alternatives, description, example, default, required, direction, multiple, multiple_sep, `type`)
+    copy(name, alternatives, description, example, default, required, min, max, direction, multiple, multiple_sep, `type`)
   }
 }